cfgmgr32: Split sources to separate files.
(cherry picked from commit d5c0e9b08a070d59bbbf0220182004c1e6b5327e)
