+++ dlls/cfgmgr32/Makefile.in
@@ -3,4 +3,6 @@ IMPORTLIB = cfgmgr32
 IMPORTS   = advapi32 rpcrt4 sechost setupapi
 
 SOURCES = \
-	main.c
+	cfgmgr32.c \
+	devobject.c \
+	notification.c
